功能修改

+++ bs4/baidu/mm131.py
@@ -7,13 +7,18 @@ import ssl
 import requests
 import os
 
-host_name = "http://www.mm131.com/"
-key_names = ["xinggan", "qingchun", "xiaohua", "chemo", "qipao", "mingxing"]
-# key_names = ["chemo", "qipao", "mingxing"]
-# key_names = ["xiaohua"]
 
 import random
 
+
+
+
+root_dir = "H:/开发视频/temp/mm131"
+host_name = "http://www.mm131.com/"
+key_names = ["xinggan", "qingchun", "xiaohua", "chemo", "qipao", "mingxing"]
+# key_names = ["性感美女", "qingchun", "xiaohua", "chemo", "qipao", "mingxing"]
+# key_names = ["chemo", "qipao", "mingxing"]
+# key_names = ["xiaohua"]
 ssl._create_default_https_context = ssl._create_unverified_context
 uapools = [
     "Mozilla/4.0 (compatible; MSIE 6.0; Windows NT 5.1; SV1; AcooBrowser; .NET CLR 1.1.4322; .NET CLR 2.0.50727)",
@@ -35,6 +40,8 @@ uapools = [
 ]
 
 
+
+
 def getDataByUrl(url):
     ua(uapools)
     data = urllib.request.urlopen(url).read().decode("gb2312", "ignore")
@@ -50,31 +57,6 @@ def ua(uapools):
     urllib.request.install_opener(opener)
 
 
-def getDataByKey(key_names):
-    for keyname in key_names:
-        for index in range(3):  # 每个大的类别获取3页的套图，每页现在是20组，总共每个大的类别获取60组
-            url = host_name + keyname + "/"
-            if index != 0:
-                if keyname == "xinggan":
-                    url += "list_6_" + str(index + 1) + ".html"
-                elif keyname == "qingchun":
-                    url += "list_1_" + str(index + 1) + ".html"
-                elif keyname == "xiaohua":
-                    url += "list_2_" + str(index + 1) + ".html"
-                elif keyname == "chemo":
-                    url += "list_3_" + str(index + 1) + ".html"
-                elif keyname == "qipao":
-                    url += "list_4_" + str(index + 1) + ".html"
-                elif keyname == "mingxing":
-                    url += "list_5_" + str(index + 1) + ".html"
-
-            # print("url-->" + url)
-            data = getDataByUrl(url)
-            pat_link = '<dd><a target="_blank" href="(.*?)"><img src='
-            pat_title = 'alt="(.*?)"\s{0,}width="'
-            linkArray = re.compile(pat_link, re.S).findall(data)  # 获取链接的正则
-            # titleArray = re.compile(pat_title, re.S).findall(data)  # 获取标题的正则
-            getDataByLink(linkArray, keyname)
 
 
 def getDataByLink(linkArray, keyName):
@@ -90,12 +72,12 @@ def getDataByLink(linkArray, keyName):
         getDataByGroup(link, pageCounts, keyName, title[0])
 
 
-def getImgUrlByLink(link, keyName, title, pageIndex):
+def getImgUrlByLink(link, keyName, title,pageCount, pageIndex):
     # print("每个高清图片所在的链接地址--->" + link)
     data = getDataByUrl(link)
     pat_img = '<div class="content-pic">\s{0,}<a href=.{1,}src="(.*?)" /></a></div>'
     imgUrl = re.compile(pat_img, re.S).findall(data)
-    saveImgToLocal(imgUrl[0], keyName, title, pageIndex)
+    saveImgToLocal(imgUrl[0], keyName, title,pageCount,pageIndex)
     # print(imgUrl)
 
 
@@ -103,14 +85,14 @@ def getImgUrlByLink(link, keyName, title, pageIndex):
 def getDataByGroup(link, pageCounts, keyName, title):
     for pageCount in pageCounts:
         # pageCount代表每个套图中有多少张高清图片
-        # print("每套套图的高清图片数量:" + pageCount)
+        print("每套套图的高清图片数量:" + pageCount)
         for i in range(int(pageCount)):
             if i != 0:
                 # 如果不是首页的话，就开始替换字符串～
                 newLink = link.replace(".html", "_" + str(i + 1) + ".html")
-                getImgUrlByLink(newLink, keyName, title, i)
+                getImgUrlByLink(newLink, keyName, title,pageCount,i)
             else:
-                getImgUrlByLink(link, keyName, title, i)
+                getImgUrlByLink(link, keyName, title,pageCount,i)
                 pass
             pass
 
@@ -119,32 +101,69 @@ def getDataByGroup(link, pageCounts, keyName, title):
 # keyName: 大的类别关键字，xinggan,qingchun,xiaohua
 # title: 每个套图的标题
 # pageIndex: 每个套图中高保真图所在的索引号
-def saveImgToLocal(imgUrl, keyName, title, pageIndex):
-    path = "mm/" + keyName + "/" + title + "/"
+def saveImgToLocal(imgUrl, keyName, title,pageCount,pageIndex):
+    path = root_dir + "/" + keyName + "/" + title + "/"
     if not os.path.isdir(path):
         os.makedirs(path)
     path += title + str(pageIndex + 1) + '.jpg'
-    try:
-        headers = {'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8',
-                   'Accept-Encoding': 'gzip, deflate',
-                   'Accept-Language': 'zh-CN,zh;q=0.8',
-                   'Cache-Control': 'no-cache',
-                   'Connection': 'keep-alive',
-                   'Cookie': 'UM_distinctid=15fa02251e679e-05c01fdf7965e7-5848211c-144000-15fa02251e7800; bdshare_firstime=1510220189357; CNZZDATA1263415983=1653134122-1510216223-null%7C1510216223; CNZZDATA3866066=cnzz_eid%3D376479854-1494676185-%26ntime%3D1494676185; Hm_lvt_9a737a8572f89206db6e9c301695b55a=1510220189; Hm_lpvt_9a737a8572f89206db6e9c301695b55a=1510220990',
-                   'Host': 'img1.mm131.me',
-                   'Pragma': 'no-cache',
-                   'Referer': 'http://www.mm131.com/' + keyName + "/",
-                   'User-Agent': random.choice(uapools)}
-        pic = requests.get(imgUrl, headers=headers, timeout=10)
-        # urllib.request.urlretrieve(imgUrl, path)
-        print(keyName + '下的' + title + '中的第' + str(pageIndex + 1) + '张图片下载完成-->' + imgUrl)
-    except requests.exceptions.ConnectionError:
-        print(keyName + '下的' + title + '中的第' + str(pageIndex + 1) + '张图片下载【错误】------------')
+    # print(path)
+    if os.path.exists(path):
+        print(keyName + '下的' + title + '中的第' + str(pageCount) + "->" + str(pageIndex + 1) + '图片已经存在不下载-->' + imgUrl)
+    else:
+        try:
+            headers = {
+                'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8',
+                'Accept-Encoding': 'gzip, deflate',
+                'Accept-Language': 'zh-CN,zh;q=0.8',
+                'Cache-Control': 'no-cache',
+                'Connection': 'keep-alive',
+                'Cookie': 'UM_distinctid=15fa02251e679e-05c01fdf7965e7-5848211c-144000-15fa02251e7800; bdshare_firstime=1510220189357; CNZZDATA1263415983=1653134122-1510216223-null%7C1510216223; CNZZDATA3866066=cnzz_eid%3D376479854-1494676185-%26ntime%3D1494676185; Hm_lvt_9a737a8572f89206db6e9c301695b55a=1510220189; Hm_lpvt_9a737a8572f89206db6e9c301695b55a=1510220990',
+                'Host': 'img1.mm131.me',
+                'Pragma': 'no-cache',
+                'Referer': 'http://www.mm131.com/' + keyName + "/",
+                'User-Agent': random.choice(uapools)}
+            pic = requests.get(imgUrl, headers=headers, timeout=10)
+            # urllib.request.urlretrieve(imgUrl, path)
+            print(keyName + ' 下的 ' + title + ' 中的第' + str(pageCount) + "->" + str(pageIndex + 1) + '张图片下载完成-->' + imgUrl)
+        except requests.exceptions.ConnectionError:
+            print(keyName + '下的' + title + '中的第' + str(pageCount) + "->" +  str(pageIndex + 1) + '张图片下载【错误】------------')
+        fp = open(path, 'wb')
+        fp.write(pic.content)
+        fp.close()
 
-    fp = open(path, 'wb')
-    fp.write(pic.content)
-    fp.close()
 
+
+
+
+def getDataByKey(key_names):
+    for keyname in key_names:
+        for index in range(3):  # 每个大的类别获取3页的套图，每页现在是20组，总共每个大的类别获取60组
+            url = host_name + keyname + "/"
+            print(str(index),url)
+            if index != 0:
+                if keyname == "xinggan":
+                    url += "list_6_" + str(index + 1) + ".html"
+                elif keyname == "qingchun":
+                    url += "list_1_" + str(index + 1) + ".html"
+                elif keyname == "xiaohua":
+                    url += "list_2_" + str(index + 1) + ".html"
+                elif keyname == "chemo":
+                    url += "list_3_" + str(index + 1) + ".html"
+                elif keyname == "qipao":
+                    url += "list_4_" + str(index + 1) + ".html"
+                elif keyname == "mingxing":
+                    url += "list_5_" + str(index + 1) + ".html"
+
+            print("url-->" + url)
+            data = getDataByUrl(url)
+            pat_link = '<dd><a target="_blank" href="(.*?)"><img src='
+            pat_title = 'alt="(.*?)"\s{0,}width="'
+
+            linkArray = re.compile(pat_link, re.S).findall(data)  # 获取链接的正则
+            # print(linkArray)
+            titleArray = re.compile(pat_title, re.S).findall(data)  # 获取标题的正则
+            # print(titleArray)
+            getDataByLink(linkArray, keyname)
 
 # print(getDataByUrl("http://www.mm131.com/chemo/1594_4.html"))
 getDataByKey(key_names)
